Extract in dedicated module and remove platform dependency
